% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohclust.R
\name{geocutree}
\alias{geocutree}
\title{Cut a Geograpĥic Tree into Groups of Data and return an sf data.frame}
\usage{
geocutree(tree, k = NULL, h = NULL)
}
\arguments{
\item{tree}{a tree as produced by geohclust. cutree() only expects a list with components merge, height, and labels, of appropriate content each.}

\item{k}{an integer scalar or vector with the desired number of groups}

\item{h}{numeric scalar or vector with heights where the tree should be cut.
At least one of k or h must be specified, k overrides h if both are given.}
}
\value{
an \code{\link{sf::sf}} like object
}
\description{
Cuts a tree, e.g., as resulting from geohclust_poly, into several groups either by specifying the desired number(s) of groups or the cut height(s).
}
