% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohclust.R
\name{geohclust_knn}
\alias{geohclust_knn}
\title{Hierarchical clustering with contiguity constraints for point data with knn links}
\usage{
geohclust_knn(df, k = 3, method = "ward", scaling = "raw")
}
\arguments{
\item{df}{an \code{\link{sf::sf}} data.frame with polygons like features}

\item{k}{number of nearest neighbors to take for building the graph (the graph will be symmetric so some points may have in fine more neighbors)}

\item{method}{linkage criterion in ward (default) or average, median}

\item{scaling}{default scaling of the features in zscore (default) or raw (i.e. no scaling)}
}
\value{
an \code{\link{stats::hclust}} like object with additional slots
\describe{
  \item{data}{The numeric data (eventually scaled) used for the clustering}
  \item{centers}{The protoypes of each tree nodes}
  \item{leafs_geometry}{geometries of the dendrogram leafs as an sfc list}
  \item{geotree}{geometries of the dendrogram no-leafs node as an sfc list}
}
}
\description{
This function take a data.frame and performs hierarchical clustering with contiguity constraints.
}
