% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohclust.R
\name{geohclust_temp}
\alias{geohclust_temp}
\title{Hierarchical clustering with contiguity constraints for temporal data}
\usage{
geohclust_temp(df, method = "ward", scaling = "raw")
}
\arguments{
\item{df}{an \code{\link{sf::sf}} data.frame with polygons like features}

\item{method}{linkage criterion in ward (default) or average, median}

\item{scaling}{default scaling of the features in zscore (default) or raw (i.e. no scaling)}
}
\value{
an \code{\link{stats::hclust}} like object with additional slots
\describe{
  \item{data}{The numeric data (eventually scaled) used for the clustering}
  \item{centers}{The protoypes of each tree nodes}
}
}
\description{
This function take a data.frame and performs hierarchical clustering with contiguity constraints.
}
