% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohclust.R
\name{geohclust_poly}
\alias{geohclust_poly}
\title{Hierarchical clustering with contiguity constraints between polygons}
\usage{
geohclust_poly(df, method = "ward", scaling = "raw")
}
\arguments{
\item{df}{an \code{\link{sf:sf}} data.frame with polygons like features}

\item{method}{linkage criterion in ward (default) or average, median}

\item{scaling}{default scaling of the features in zscore (default) or raw (i.e. no scaling)}
}
\value{
an \code{\link{hclust::hclust}} like object
}
\description{
This function take an \code{\link{sf:sf}} data.frame and perfomrs hierarchical clustering with contiguity constraints.
}
